SCD-BC/requirements.txt
numpy
keras
tensorflow-cpu
pandas
matplotlib
pillow
flask
seaborn
gunicorn
